requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
tiktoken
pypdf
langchain
langchain-community
langchain-google-genai
langchain-openai
langchain-postgres